gvc gvrender_usershape: take 'n' as a 'size_t'
Pushes `size_t` work down the call stack.

+++ lib/common/shapes.c
@@ -3024,7 +3024,7 @@ static void poly_gencode(GVJ_t * job, node_t * n)
 		gvrender_polygon(job, AF, sides, filled);
 	    }
 	}
-	gvrender_usershape(job, name, AF, (int)sides, filled != 0,
+	gvrender_usershape(job, name, AF, sides, filled != 0,
 			   late_string(n, N_imagescale, "false"),
 			   late_string(n, N_imagepos, "mc"));
 	filled = 0;		/* with user shapes, we have done the fill if needed */
